replace gme and bitcoin links to new version
discord support has informed me that they will not change bot usernames after a bot is verified. This means that the bots for BTC - $54030  and GME - $210.0 will not ever stop saying the wrong (old) price in the title. The best solution I have is to create new version of those bots, and replace all links to the old bot's client id (invite links) to point to the new bots. If you want the old price out of your servers just kick the old bot and invite the new one. 

I will keep the old versions running as long as the rest of the bots are, so there will be no forceful removal from any servers.